Added MediaBrowserVC, updated copyright notice.
The media browser is just the camera picker with different enumeration choice right now.
Need to add another stage in the alert process to ask the user whether they want to start a new
palette or add to an old one.
Copyright notice on all files updated to include Kevin.